ci: upload code coverage in a separate step
Wait for all tests to complete before uploading code coverage.
Previously the unit tests would upload their results in advance and the
codecov bot would report coverage issues in new PRs.

As it is, the workflow will not upload code coverage reports if any of
the unit or integration tests fail. Which I think is fine.